tests/test_generation.py
import unittest
from src import generation

class TestGeneration(unittest.TestCase):
    def test_generate_script(self):
        script = generation.generate_script("Write an intro for a video.")
        self.assertTrue(len(script) > 0)

    def test_smart_clip_video(self):
        generation.smart_clip_video("sample_video.mp4", 10, 20, "clipped_video.mp4")
        # Additional checks can be added to verify the output

    def test_auto_generate_video(self):
        generation.auto_generate_video("AI in Video Editing", ["Introduction", "AI Basics", "Advanced Techniques"], "auto_generated_video.mp4")
        # Additional checks can be added to verify the output

    def test_dynamic_music_generation(self):
        generation.dynamic_music_generation("auto_generated_video.mp4", "music_track.mp3")
        # Additional checks can be added to verify the output
